Began re-implementing drawing...
Need to determine best layout for MVC solution. GraphicsScene is acting as View + Controller, but when addt'l tabs are added to the View for editing e.g. tile maps etc, each tab's GraphicsScene will need access to the same Model - in this case GameData. In C/C++ this would simply be a pass-by-reference solution, however there's the question of atomicity as well of the Model data.