Refactored and fixed batch undo/redo logic
Final verdict on performance isn't outstanding in worst-case scenarios. E.g. entire sprite sheet is active as the subject, and filling every pixel. There are several culprits contributing to the lag, on both the canvas and undo/redo system ends, with the palette in the middle. The most obvious solution - or bandaid - would be to C DLL some of the canvas and palette code. Beyond that, some tricky optimizations or a complete refactor are the only other viable paths